refactor(auth): extract csrf seam
Move CSRF validation for the built-in-auth path to a dedicated
module:

- _CSRF_PROTECTED_METHODS (frozen set of mutating HTTP methods)
- validate_csrf (header-first, form-body-fallback; constant-time
  comparison against the signed session cookie's CSRF payload)

The proxy-mode CSRF helper (``_validate_proxy_csrf``) stays in
__init__.py until Phase 2f because it lives next to its sibling
proxy-auth primitives.  tests/test_huntarr_vulns.py's
``inspect.getsource(_auth_module.validate_csrf)`` check still
reaches the function through the re-export, so the constant-time
assertion is green.
